add pixi configuration
This adds a pixi project file that specifies the kiara dependencies
that are required to run a backend, and also export Python/Pydantic
models to typescript. The commands that can be run are:

- pixi run show-version  # show installed kiara and plugin versions
- pixi run backend       # run the kiara openapi rest service
- pixi export-models     # export pytantic to typescript models

The typescript models will be exported to: "src/lib/kiara_models.ts".

"pixi" can be downloaded from: https://github.com/prefix-dev/pixi