add utils module from lumy prototype
The main thing this provides is a wrapper class around the KiaraAPI.
This should make it easy to create a client in JS, independent of which
backend protocol we choose to implement later. For now, it supports
the OpenAPI one.

url of lumy prototype: https://github.com/makkus/lumy

This module might need some updates before its functional again,
have not tested it for a while.